+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import dst
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  DstPlugin.register(with: registry.registrar(forPlugin: "DstPlugin"))
+}